Version 2: Redesigned encryption, improved speed & more!
This new version of Firecrypt ships with the following changes:
- Encryption algorithm changed from AES-CBC to XChaCha20-Poly1305
- Keys are now derived with Argon2
- Major speed improvements
- New icon
- Profiles are now inaccessible from Firefox while being configured
- Dependency updates
And more!

As a result of the completely overhauled encryption system,
old profiles will need to be migrated. This version of Firecrypt
does support migrating legacy profiles, however future versions may not.
To migrate a locked profile, simply select it from the profile list,
enter the password, and agree to the migration. You can migrate an unlocked
profile by just clicking it and choosing a new password. The old files
will be automatically removed.